Update API Consumption -RESt and databese

+++ Gestion_Capacitaciones/app/src/main/java/com/ista/gestion_capacitaciones/db/DbUsuarios.java
@@ -22,7 +22,7 @@ public class DbUsuarios extends SqlConexion{
         this.context=context;
     }
 
-    public long insertaUsuario(String username, String password, int personaId) {
+    public long insertaUsuario(String username, String password, Long personaId) {
         SQLiteDatabase db = this.getWritableDatabase();
         long result = -1;
         try{
